NumberZap: delete code number of digits
Make number zapping easier. Do not set a parameter for a maximum number of digits, the user must know that this parameter exists.
e.g. If you have a number less five digits set and you zap to a larger number, you simply won't see the last number you set on the Zap on the skin screen.
Let the screen display all user input in numbers and tell the user when they set a number that the service does not exist (new information: widget service name). There are no limits on access to the number.
follow 9a482e2