Add possibility that plugins can add instantiate screens to the InfoBars
With this commit it is possible that e.g. Weather plugins, but also any
other plugin is able to add on demand a screen to the InfoBar and/or
SecondInfobar. So for these types of plugins simply screens can be added
and shown in addition to the infobar. So no need to add custom renderers
for these type of adds.

Therefore I added 2 types of plugins:

        WHERE_INFOBAR_SCREEN = 16
        WHERE_SECONDINFOBAR_SCREEN = 17

In the plugin itself should return the specific screen.

So you need to return a plugin descrypter like as in the example below:

class DummyScreen(Screen):
	def __init__(self, session, **kwargs):
		self.skin = """<screen position="0,0" size="200,200" zPosition="200"/>"""
		Screen.__init__(self, session)

def Plugins(**kwargs):
	return [PluginDescriptor(where=[PluginDescriptor.WHERE_INFOBAR_SCREEN], fnc=DummyScreen),
		PluginDescriptor(where=[PluginDescriptor.WHERE_SECONDINFOBAR_SCREEN], fnc=DummyScreen)]