[keymap.xml] Add new ShowVirtualKeyboardActions action map
This is a new action map to allow for a future clarification that will make all VirtualKeyboard actions use a standardised case.

Having two separate action maps that differ by the case of one letter that do very different things is not a good idea and can cause confusion.

Co-Authored-By:
IanSav <
IanSav@gmx.net>